// angular-frontend/src/models/bike.model.ts
export interface Bike {
  bikeModel: string | null | undefined;
  parentCompany: string | null | undefined;
  launchYear: string | null | undefined;
  price: string | null | undefined;
  fuelType: string | null | undefined;
  bikeType: string | null | undefined;
  cubicCapacity: string | null | undefined;
  weight: string | null | undefined;
  torque: string | null | undefined;
  horsePower: string | null | undefined;
  mileage: string | null | undefined;
  gears: string | null | undefined;
}